+++ modules/features/widgets/src/main/kotlin/au/com/shiftyjelly/pocketcasts/widget/ui/SkipForwardButton.kt
@@ -2,17 +2,15 @@ package au.com.shiftyjelly.pocketcasts.widget.ui
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.unit.dp
 import androidx.glance.ColorFilter
 import androidx.glance.GlanceModifier
-import androidx.glance.GlanceTheme
 import androidx.glance.Image
 import androidx.glance.ImageProvider
 import androidx.glance.LocalContext
 import androidx.glance.action.clickable
-import androidx.glance.layout.Box
+import androidx.glance.layout.Alignment
+import androidx.glance.layout.fillMaxSize
 import androidx.glance.layout.padding
-import androidx.glance.layout.size
 import androidx.glance.semantics.contentDescription
 import androidx.glance.semantics.semantics
 import au.com.shiftyjelly.pocketcasts.widget.action.SkipForwardAction
@@ -22,27 +20,27 @@ import au.com.shiftyjelly.pocketcasts.localization.R as LR
 
 @Composable
 internal fun SkipForwardButton(
-    size: Dp = 36.dp,
+    iconPadding: Dp,
+    modifier: GlanceModifier = GlanceModifier,
+    isClickable: Boolean = true,
 ) {
     val contentDescription = LocalContext.current.getString(LR.string.skip_forward)
+    val source = LocalSource.current
 
-    Box(
-        modifier = GlanceModifier
-            .size(size)
-            .clickable(SkipForwardAction.action(LocalSource.current))
-            .semantics { this.contentDescription = contentDescription },
+    RounderCornerBox(
+        contentAlignment = Alignment.Center,
+        backgroundTint = LocalWidgetTheme.current.buttonBackground,
+        modifier = modifier.applyIf(isClickable) { ifModifier ->
+            ifModifier
+                .clickable(SkipForwardAction.action(source))
+                .semantics { this.contentDescription = contentDescription }
+        },
     ) {
-        Image(
-            provider = ImageProvider(IR.drawable.ic_circle),
-            contentDescription = null,
-            colorFilter = ColorFilter.tint(GlanceTheme.colors.primary),
-            modifier = GlanceModifier.size(size),
-        )
         Image(
             provider = ImageProvider(IR.drawable.ic_widget_skip_forward),
             contentDescription = null,
-            colorFilter = ColorFilter.tint(GlanceTheme.colors.onPrimary),
-            modifier = GlanceModifier.size(size).padding(size / 5),
+            colorFilter = ColorFilter.tint(LocalWidgetTheme.current.icon),
+            modifier = GlanceModifier.fillMaxSize().padding(vertical = iconPadding),
         )
     }
 }